requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-crispy-forms==1.9.2
django-grappelli==2.14.2
django-heroku==0.3.1
django-tinymce4-lite==1.8.0
gunicorn==20.0.4
jsmin==2.2.2
Pillow==7.2.0
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
